pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quick-core"
version = "0.0.0rc"
dependencies = [
    "cirq-core == 1.4.1",
    "genQC == 0.1.0",
    "numpy >= 1.23,< 3.0",
    "pennylane == 0.39.0",
    "pytket == 1.37.0",
    "pytket-qiskit == 0.62.0",
    "pytket-cirq == 0.39.0",
    "qiskit == 1.3.1",
    "qiskit_aer == 0.16.0",
    "qiskit_ibm_runtime == 0.34.0",
    "qiskit-transpiler-service == 0.4.14",
    "quimb == 1.10.0",
    "tket2 == 0.6.0"
]
requires-python = ">=3.10, <3.13"
authors = [
  {name = "Amir Ali Malekani Nezhad", email = "amiralimlk07@gmail.com"},
]
maintainers = [
  {name = "Amir Ali Malekani Nezhad", email = "amiralimlk07@gmail.com"}
]
description = "quick is an agnostic gate-based circuit SDK, providing an integrated interface for using any supported quantum circuit framework seamlessly."
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["quantum computing", "quantum circuit", "quantum compiler"]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent"
]

[project.optional-dependencies]
dev = [
  "ruff == 0.9.2",
  "mypy == 1.14.1",
  "pytest == 8.3.4",
  "pytest-mock == 3.14.0",
  "pytest-cov == 6.0.0",
  "coverage == 7.6.10",
]

[project.urls]
Repository = "https://github.com/qualition/quick.git"
"Bug Tracker" = "https://github.com/qualition/quick/issues"

[tool.setuptools.packages.find]
include = ["quick*"]

[tool.codeflash]
# All paths are relative to this pyproject.toml's directory.
module-root = "quick"
tests-root = "tests"
test-framework = "pytest"
ignore-paths = []
formatter-cmds = ["disabled"]
